# backend-app/docs/openapi.yaml
openapi: 3.0.0
info:
  title: FlexiCharge API
  version: 1.0.0
servers:
  - url: http://localhost:8080/
    description: Development Environment
  - url: http://18.202.253.30:8080/
    description: Production Environment (AWS)
paths:
...
  /transactions:
    post:
      tags:
        - Transactions
      summary: Create a transaction
      description: |-
        Creates a transaction that can be started using the `transactionID` returned
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                chargerID:
                  type: integer
                  example: 100000
                isKlarnaPayment:
                  type: boolean
                  example: true
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Error message
        '500':
          description: Error message
  /transactions/start/{transactionID}:
    parameters:
      - in: path
        name: transactionID
        schema:
          type: integer
          example: 100000
        required: true
    put:
      tags:
        - Transactions
      summary: Start a transaction
      description: |-
        Starts a transaction
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Error message
        '500':
          description: Error message
  /transactions/{transactionID}:
    parameters:
      - in: path
        name: transactionID
        schema:
          type: integer
          example: 100000
        required: true
    get:
      tags:
        - Transactions
      summary: Get status of a transaction
      description: |-
        Gets the status of a transaction
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          description: Transaction not found
        '500':
          description: Error message
  /transactions/stop/{transactionID}:
    parameters:
      - in: path
        name: transactionID
        schema:
          type: integer
          example: 100000
        required: true
    put:
      tags:
        - Transactions
      summary: Stop a transaction
      description: |-
        Stops a transaction
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Error message
        '500':
          description: Error message
  /transactions/userTransactions/{userID}:
    parameters:
      - in: path
        name: userID
        schema:
          type: integer
        required: true
    get:
      tags:
        - Transactions
      summary: Transaction by userID
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
  /transactions/chargerTransactions/{chargerID}:
    parameters:
      - in: path
        name: chargerID
        schema:
          type: integer
        required: true
    get:
      tags:
        - Transactions
      summary: Transaction by transactions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
...
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        transactionID:
          type: integer
          example: 9999
        isKlarnaPayment:
          type: boolean
          example: false
        kwhTransfered:
          type: number
          example: 50
        currentChargePercentage:
          type: number
          example: 79
        pricePerKwh:
          type: string
          example: '123.00'
        timestamp:
          type: integer
          example: 1663663253
        paymentID:
          type: string
          example: null
        userID:
          type: string
          example: 'xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx'
        session_id:
          type: string
          example: null
        klarna_consumer_token:
          type: string
          example: null
        paymentConfirmed:
          type: boolean
          example: null
        meterStart:
          type: integer
          example: 1
        chargerID:
          type: integer
          example: 100000
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer